Massive animation update! Update navigation library to alpha 6, introduced shared transitions, fixed a few bugs (empty url on dialog not showing error, coroutines scope doesn't work fine with dialogs, visibility button on text diff is gone (hopefully forever)). Also, WorkManager was updated and should fix additional bugs. Everything is more fluid and faster than ever before.

+++ app/src/main/java/com/bernaferrari/changedetection/detailstext/TextViewModel.kt
@@ -38,8 +38,8 @@ class TextViewModel(
      *
      * @param id The diff url to be removed.
      */
-    fun removeSnap(id: String) = GlobalScope.launch {
-        mSnapsRepository.deleteSnap(id)
+    fun removeSnap(id: String?) = GlobalScope.launch {
+        if (id != null) mSnapsRepository.deleteSnap(id)
     }
 
     private var currentJob: Job? = null
